# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OFFITEC ERP API",
    "version": "1.0.0",
    "description": "OFFITEC Kurumsal ERP Sistemi API Dokümantasyonu"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/articles": {
      "get": {
        "tags": [
          "Articles"
        ],
        "summary": "Ürün/Malzeme listesini getir (search, kategori, stok özetli)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "onlyActive",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "includeStock",
            "schema": {
              "type": "boolean",
              "description": "Stok bakiyelerini de döndür"
            }
          }
        ]
      },
      "post": {
        "tags": [
          "Articles"
        ],
        "summary": "Yeni ürün/malzeme oluştur",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/articles/lookup/{code}": {
      "get": {
        "tags": [
          "Articles"
        ],
        "summary": "Stok kodu veya barkod ile ürünü getir (Barkod tarayıcı)",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/articles/{id}": {
      "get": {
        "tags": [
          "Articles"
        ],
        "summary": "Ürün detayı",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Articles"
        ],
        "summary": "Ürünü güncelle",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Articles"
        ],
        "summary": "Ürünü sil",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/attendance/check-in": {
      "post": {
        "tags": [
          "Attendance"
        ],
        "summary": "Personel giriş kaydı oluştur",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Giriş kaydedildi"
          },
          "400": {
            "description": "Hata (zaten giriş yapılmış vb.)"
          }
        }
      }
    },
    "/attendance/check-out": {
      "post": {
        "tags": [
          "Attendance"
        ],
        "summary": "Personel çıkış kaydı oluştur",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Çıkış kaydedildi"
          },
          "400": {
            "description": "Hata (giriş bulunamadı vb.)"
          }
        }
      }
    },
    "/attendance": {
      "get": {
        "tags": [
          "Attendance"
        ],
        "summary": "Devam kayıtlarını listele (tarih bazlı)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Tarih filtresi (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Devam listesi"
          },
          "401": {
            "description": "Yetkisiz"
          }
        }
      }
    },
    "/attendance/{id}": {
      "patch": {
        "tags": [
          "Attendance"
        ],
        "summary": "Devam kaydını manuel düzelt (İK yetkisi)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "checkInTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "checkOutTime": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Kayıt güncellendi (audit loglu)"
          },
          "400": {
            "description": "Hata"
          }
        }
      }
    },
    "/attendance/me/today": {
      "get": {
        "tags": [
          "Attendance"
        ],
        "summary": "Giriş yapan kullanıcının bugünkü giriş-çıkış bilgisini getir",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/tenders/{id}/positions/{positionId}/articles": {
      "post": {
        "tags": [
          "Tender"
        ],
        "summary": "İhale pozisyonuna dahili ürün (BOM) bağlar ve malzeme maliyetini otomatik hesaplar.",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}